refactor: make CLI parse public
This is in order to make the testing easier. There is no need for users to use the CLI class anyways. The public Parse might be a bit confusing but there is no harm in it being public as it only parses the arguemnts. In the CLI, this parse function is used in the constructor so in the CLI executable I can still just use it the same way I do now.