fix: update packages to latest version
This also includes updating major versions for some packages.

In addition to this update the import of yargs in the CLI has been changed to get rid of a yargs warning message. See https://github.com/yargs/yargs/issues/1439 for info on that. The CLI tests has an added ignoreline for properly disabling the lint warnings for mocking the `process.exit` implementation.