feat(cli): add possibility to pass files to lint
This is used in combination with the existing way of passing a glob pattern. This adds support for a list of glob patterns and/or files (absolute and/or relative), more or less the same way ESLint gets its files.

WIP: This is a work in progress and is currently functional but does not ignore files.